requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.3
django-cors-headers==4.6.0
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.6
PyJWT==2.9.0
sqlparse==0.5.2
tzdata==2024.2
